allow rolling back SLO targets
